rename components for clarity and introducing click outside handler to close down the login dropdown
